+++ riverdriver/riverdrivertest/sql_fragments.go
@@ -14,6 +14,59 @@ import (
 func exerciseSQLFragments[TTx any](ctx context.Context, t *testing.T, executorWithTx func(ctx context.Context, t *testing.T) (riverdriver.Executor, riverdriver.Driver[TTx])) {
 	t.Helper()
 
+	t.Run("SQLFragmentColumnContainsAll", func(t *testing.T) {
+		t.Parallel()
+
+		exec, driver := executorWithTx(ctx, t)
+
+		var (
+			job1 = testfactory.Job(ctx, t, exec, &testfactory.JobOpts{Tags: []string{"alpha", "beta"}})
+			_    = testfactory.Job(ctx, t, exec, &testfactory.JobOpts{Tags: []string{"alpha"}})
+			_    = testfactory.Job(ctx, t, exec, &testfactory.JobOpts{Tags: []string{"beta"}})
+		)
+
+		const namedArg = "tags_all"
+		sqlFragment, arg, err := driver.SQLFragmentColumnContainsAll("tags", namedArg, []string{"alpha", "beta"})
+		require.NoError(t, err)
+
+		jobs, err := exec.JobList(ctx, &riverdriver.JobListParams{
+			Max:           100,
+			NamedArgs:     map[string]any{namedArg: arg},
+			OrderByClause: "id",
+			WhereClause:   sqlFragment,
+		})
+		require.NoError(t, err)
+		require.Len(t, jobs, 1)
+		require.Equal(t, job1.ID, jobs[0].ID)
+	})
+
+	t.Run("SQLFragmentColumnContainsAny", func(t *testing.T) {
+		t.Parallel()
+
+		exec, driver := executorWithTx(ctx, t)
+
+		var (
+			job1 = testfactory.Job(ctx, t, exec, &testfactory.JobOpts{Tags: []string{"alpha"}})
+			job2 = testfactory.Job(ctx, t, exec, &testfactory.JobOpts{Tags: []string{"beta"}})
+			_    = testfactory.Job(ctx, t, exec, &testfactory.JobOpts{Tags: []string{"gamma"}})
+		)
+
+		const namedArg = "tags_any"
+		sqlFragment, arg, err := driver.SQLFragmentColumnContainsAny("tags", namedArg, []string{"alpha", "beta"})
+		require.NoError(t, err)
+
+		jobs, err := exec.JobList(ctx, &riverdriver.JobListParams{
+			Max:           100,
+			NamedArgs:     map[string]any{namedArg: arg},
+			OrderByClause: "id",
+			WhereClause:   sqlFragment,
+		})
+		require.NoError(t, err)
+		require.Len(t, jobs, 2)
+		require.Equal(t, job1.ID, jobs[0].ID)
+		require.Equal(t, job2.ID, jobs[1].ID)
+	})
+
 	t.Run("SQLFragmentColumnIn", func(t *testing.T) {
 		t.Parallel()