+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added `JobListParams.Tags` for filtering jobs by one or more case-insensitive tags. [PR #1339](https://github.com/riverqueue/river/pull/1339).
+- Added `JobListParams.TagsAll` and `JobListParams.TagsAny` for filtering jobs that match every or any exact tag, respectively. [PR #1339](https://github.com/riverqueue/river/pull/1339).
 
 ## [0.42.0] - 2026-07-31